rename loaders package to loader
