+++ v2/loader/doc.go
@@ -17,4 +17,4 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package loaders // import "go.mercari.io/yo/v2/loaders"
+package loader // import "go.mercari.io/yo/v2/loader"
